requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==1.3.0
    # via
    #   chex
    #   jax
    #   jaxlib
    #   jaxopt
    #   optax
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
appdirs==1.4.4
    # via yfinance
beautifulsoup4==4.11.1
    # via yahoofinancials
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
chex==0.1.5
    # via optax
contourpy==1.0.6
    # via matplotlib
cvxpy==1.2.2
    # via -r requirements.in
cycler==0.11.0
    # via matplotlib
dm-tree==0.1.7
    # via chex
ecos==2.0.10
    # via cvxpy
flatbuffers==2.0.7
    # via jaxlib
fonttools==4.38.0
    # via matplotlib
get-all-tickers==1.8
    # via -r requirements.in
hydra-core==1.2.0
    # via -r requirements.in
idna==3.4
    # via requests
inflection==0.5.1
    # via quandl
jax==0.3.1
    # via
    #   -r requirements.in
    #   chex
    #   jaxopt
    #   optax
jaxlib==0.3.0
    # via
    #   -r requirements.in
    #   chex
    #   jaxopt
    #   optax
jaxopt==0.4
    # via -r requirements.in
joblib==1.2.0
    # via scikit-learn
kiwisolver==1.4.4
    # via matplotlib
lxml==4.9.1
    # via yfinance
matplotlib==3.6.2
    # via
    #   -r requirements.in
    #   jaxopt
more-itertools==9.0.0
    # via quandl
multitasking==0.0.11
    # via yfinance
numpy==1.23.5
    # via
    #   -r requirements.in
    #   chex
    #   contourpy
    #   cvxpy
    #   ecos
    #   jax
    #   jaxlib
    #   jaxopt
    #   matplotlib
    #   opt-einsum
    #   optax
    #   osqp
    #   pandas
    #   quandl
    #   scikit-learn
    #   scipy
    #   scs
    #   yfinance
omegaconf==2.2.3
    # via hydra-core
opt-einsum==3.3.0
    # via jax
optax==0.1.1
    # via -r requirements.in
osqp==0.6.2.post8
    # via cvxpy
packaging==21.3
    # via
    #   hydra-core
    #   matplotlib
pandas==1.5.2
    # via
    #   -r requirements.in
    #   get-all-tickers
    #   quandl
    #   yfinance
pillow==9.3.0
    # via matplotlib
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
    #   quandl
pytz==2022.6
    # via
    #   pandas
    #   yahoofinancials
pyyaml==6.0
    # via omegaconf
qdldl==0.1.5.post2
    # via osqp
quandl==3.7.0
    # via -r requirements.in
requests==2.28.1
    # via
    #   get-all-tickers
    #   quandl
    #   yfinance
scikit-learn==1.1.3
    # via -r requirements.in
scipy==1.9.3
    # via
    #   -r requirements.in
    #   cvxpy
    #   ecos
    #   jax
    #   jaxlib
    #   jaxopt
    #   osqp
    #   scikit-learn
    #   scs
scs==3.2.2
    # via cvxpy
six==1.16.0
    # via
    #   python-dateutil
    #   quandl
sklearn==0.0.post1
    # via -r requirements.in
soupsieve==2.3.2.post1
    # via beautifulsoup4
threadpoolctl==3.1.0
    # via scikit-learn
toolz==0.12.0
    # via chex
typing-extensions==4.4.0
    # via
    #   jax
    #   optax
urllib3==1.26.13
    # via requests
yahoofinancials==1.6
    # via -r requirements.in
yfinance==0.1.87
    # via -r requirements.in
